requirements.txt
django>=5.0,<6.0
djangorestframework>=3.14,<4.0
pillow>=10.0,<11.0
drf-spectacular>=0.28.0,<0.29.0
